Day 9a Refacor of day 1 for change in Solver int->ulong

+++ Puzzles.Tests/Day1/PuzzleDaySolver1Tests.cs
@@ -10,7 +10,7 @@ namespace Puzzles.Tests.Day1
 
         [Theory]
         [MemberData(nameof(KNumbersThatSumToN))]
-        public void Should_GetKNumbersThatSumToN(List<int> numbers, int k, int n, List<int> expected)
+        public void Should_GetKNumbersThatSumToN(List<ulong> numbers, int k, ulong n, List<ulong> expected)
         {
             var solver = new PuzzleSolverDay1();
             var result = solver.GetKNumbersThatSumToN(numbers, k, n);
@@ -20,7 +20,7 @@ namespace Puzzles.Tests.Day1
 
         [Theory]
         [MemberData(nameof(KNumbersThatNotSumToN))]
-        public void ShouldNot_GetKNumbersThatSumToN(List<int> numbers, int k, int n)
+        public void ShouldNot_GetKNumbersThatSumToN(List<ulong> numbers, int k, ulong n)
         {
             var solver = new PuzzleSolverDay1();
             var result =  solver.GetKNumbersThatSumToN(numbers, k, n);
@@ -30,7 +30,7 @@ namespace Puzzles.Tests.Day1
 
         [Theory]
         [MemberData(nameof(TooFewNumbers))]
-        public void ShouldNot_GetKNumbersThatSumToN_TooFewNumbers(List<int> numbers, int k, int n)
+        public void ShouldNot_GetKNumbersThatSumToN_TooFewNumbers(List<ulong> numbers, int k, ulong n)
         {
             var solver = new PuzzleSolverDay1();
             Assert.Throws<ArgumentException>("k", () => solver.GetKNumbersThatSumToN(numbers, k, n));
@@ -38,7 +38,7 @@ namespace Puzzles.Tests.Day1
 
         [Theory]
         [MemberData(nameof(NumbersThatSumToN))]
-        public void Should_GetNumbersThatSumToN(List<int> numbers, int n)
+        public void Should_GetNumbersThatSumToN(List<ulong> numbers, ulong n)
         {
             var solver = new PuzzleSolverDay1();
             var result = solver.GetNumbersThatSumToN(numbers, n);
@@ -49,7 +49,7 @@ namespace Puzzles.Tests.Day1
 
         [Theory]
         [MemberData(nameof(NumbersThatNotSumToN))]
-        public void ShouldNot_GetNumbersThatSumToN(List<int> numbers, int n)
+        public void ShouldNot_GetNumbersThatSumToN(List<ulong> numbers, ulong n)
         {
             var solver = new PuzzleSolverDay1();
             var result = solver.GetNumbersThatSumToN(numbers, n);
@@ -59,55 +59,55 @@ namespace Puzzles.Tests.Day1
 
         public static IEnumerable<object[]> NumbersThatSumToN()
         {
-            yield return new object[] { new List<int>() { 1, 2, 3, 4, 2017 }, 2020 };
-            yield return new object[] { new List<int>() { 2011, 300, 3, 4, 200, 1720 }, 2020 };
-            yield return new object[] { new List<int>() { 2011, 300, 3, 4, 200, 1720 }, 2311 };
+            yield return new object[] { new List<ulong>() { 1, 2, 3, 4, 2017 }, 2020 };
+            yield return new object[] { new List<ulong>() { 2011, 300, 3, 4, 200, 1720 }, 2020 };
+            yield return new object[] { new List<ulong>() { 2011, 300, 3, 4, 200, 1720 }, 2311 };
         }
         public static IEnumerable<object[]> NumbersThatNotSumToN()
         {
-            yield return new object[] { new List<int>() { 3000 }, 2020 };
-            yield return new object[] { new List<int>() { 10, 31, 50, 11, 231, 2308, 32123 }, 2020 };
+            yield return new object[] { new List<ulong>() { 3000 }, 2020 };
+            yield return new object[] { new List<ulong>() { 10, 31, 50, 11, 231, 2308, 32123 }, 2020 };
         }
 
         public static IEnumerable<object[]> KNumbersThatSumToN()
         {
             yield return new object[] {
-                new List<int>() { 1, 2, 3, 5, 6, 2017 },
+                new List<ulong>() { 1, 2, 3, 5, 6, 2017 },
                 2, 2020,
-                new List<int>() {3,2017 }
+                new List<ulong>() {3,2017 }
             };
 
             yield return new object[] {
-                new List<int>() { 20, 33, 44, 57, 199, 200 ,1800},
+                new List<ulong>() { 20, 33, 44, 57, 199, 200 ,1800},
                 3, 2020,
-                new List<int>() { 200, 1800, 20 } };
+                new List<ulong>() { 200, 1800, 20 } };
 
             yield return new object[] {
-                new List<int>() { 20, 33, 44, 57, 199, 200 ,1800},
+                new List<ulong>() { 20, 33, 44, 57, 199, 200 ,1800},
                 3 ,97,
-                new List<int>(){ 33, 44, 20 } };
+                new List<ulong>(){ 33, 44, 20 } };
         }
         public static IEnumerable<object[]> KNumbersThatNotSumToN()
         {
             yield return new object[] {
-                new List<int>() { 1, 2, 3, 5, 6, 2017 },
+                new List<ulong>() { 1, 2, 3, 5, 6, 2017 },
                 2, 2050
             };
 
             yield return new object[] {
-                new List<int>() { 20, 33, 44, 57, 199, 200 ,1800},
+                new List<ulong>() { 20, 33, 44, 57, 199, 200 ,1800},
                 7, 2020
             };
 
             yield return new object[] {
-                new List<int>() { 20, 33, 44, 57, 199, 200 ,1800},
+                new List<ulong>() { 20, 33, 44, 57, 199, 200 ,1800},
                 2 ,97
             };
         }
         public static IEnumerable<object[]> TooFewNumbers()
         {
             yield return new object[] {
-                new List<int>() { 1, 2, 3, 5, 6, 2017 },
+                new List<ulong>() { 1, 2, 3, 5, 6, 2017 },
                 7, 2020
             };
         }
